`lisp-comment-dwim-region` better handles mixed content
It can simultaneously (un)comment a mix of text & s-expressions.

When you have

  this & that
  (foo :bar)

and you select both lines and M-x lisp-comment-dwim-region, it'll turn
to:

  ; this & that
  #+nil (foo :bar)

And if you subsequently select that & run lisp-comment-dwim-region,
it'll turn it back.

But if you happen to have:

  this & that
  #+nil (foo :bar)

Then and you select and apply lisp-comment-dwim-region, it'll comment
the uncommented and uncomment the commented region.